mobile-usb-networking: create a dual RNDIS/ECM + ACM configuration
Create a Windows-compatible RNDIS6 gadget together with ECM
configuration for Linux and OSX. Linux and OSX will automatically switch
to using ECM configuration, while Windows from 7 onwards will use RNDIS6
thanks to vendor-specific descriptors.

Derive the MAC addresses assigned to the interfaces from machine-id too,
and ensure they are deterministic, so the IP addresses chosen by dnsmasq
are deterministic too, and the connection is consistently recognized by
the USB host, especially when NetowrkManager is running on it as well.

Add an ACM port to each configuration, which will expose
a login shell, useful for development or for recovery situations if no
USB OTG keyboard is available.

Finally, clean up the descripors, ensure the device is reporting
bus-powered state with 500mA current, and has correct
class/subclass/protocol for a composite device.

Signed-off-by: Lech Perczak <lech.perczak@gmail.com>